+++ client/src/shared/adminPage/components/CheckBox.js
@@ -1,0 +1,9 @@
+function CheckBox({ onChange, checked }) {
+  return (
+    <>
+      <input type="checkbox" onChange={onChange} checked={checked} />
+    </>
+  );
+}
+
+export default CheckBox;
