+++ client/src/shared/adminPage/components/DataTable.js
@@ -1,9 +1,13 @@
 import React, { useState } from "react";
 
-// import "../components/DataTable.css";
 import * as S from "./DataTableStyle.js";
-import SelectBox from "./SelectBox.js";
 import AdminButtonGrey from "../../../components/adminPage/AdminButtonGrey.js";
+
+import SelectBox from "./SelectBox.js";
+import CheckBox from "./CheckBox.js";
+import CheckBoxHandlerChecked from "./CheckBoxHandlerChecked.js";
+import CheckBoxHandlerSelectAll from "./CheckBoxHandlerSelectAll.js";
+import CheckBoxHandlerDelete from "./CheckBoxHandlerDelete.js";
 
 function DataTable({ keySet, headers, items, setItems, setDelete, btnMsg }) {
   if (!headers || !headers.length) {
@@ -13,56 +17,21 @@ function DataTable({ keySet, headers, items, setItems, setDelete, btnMsg }) {
   const [selectedStatus, setSelectedStatus] = useState("");
 
   const onChecked = (item) => {
-    const newSelectedLists = new Set(selectedLists);
-    const checkedItem = items.findIndex((i) => i === item);
-    if (newSelectedLists.has(checkedItem)) {
-      newSelectedLists.delete(checkedItem);
-    } else {
-      newSelectedLists.add(checkedItem);
-    }
-
-    const selectedStatus =
-      Array.from(newSelectedLists).length > 0 ? "your_selected_value" : "";
-    setSelectedStatus(selectedStatus);
-
-    setSelectedLists(newSelectedLists);
-
-    const selectedContents = Array.from(newSelectedLists).map(
-      (index) => items[index]
-    );
-    const selectedProductId = Array.from(newSelectedLists).map(
-      (index) => items[index].productID
-    );
-
-    console.log("selectedContents", selectedContents);
-    console.log("selectedProductId", selectedProductId);
+    CheckBoxHandlerChecked({
+      item,
+      selectedLists,
+      setSelectedLists,
+      setSelectedStatus,
+      items,
+    });
   };
 
   const SelectAll = () => {
-    const selectedAll = selectedLists.size === items.length;
-    const newSelectedLists = new Set(
-      selectedAll ? [] : items.map((_, index) => index)
-    );
-    setSelectedLists(newSelectedLists);
-
-    const selectedAllContents = Array.from(newSelectedLists).map(
-      (index) => items[index]
-    );
-
-    console.log("selectedAllContents", selectedAllContents);
+    CheckBoxHandlerSelectAll({ selectedLists, items, setSelectedLists });
   };
 
   const deleteChecked = () => {
-    const selectedIndexes = Array.from(selectedLists);
-    const updatedItems = items.filter(
-      (_, index) => !selectedIndexes.includes(index)
-    );
-    const deletedItems = selectedIndexes.map((index) => items[index]);
-    setItems(updatedItems);
-
-    console.log("삭제 전 배열", items);
-    console.log("뭐가 삭제될거니", deletedItems);
-    console.log("삭제 후 배열", updatedItems);
+    CheckBoxHandlerDelete({ selectedLists, items, setItems });
   };
 
   const headerList = headers.map((header) => header.value);
@@ -89,8 +58,7 @@ function DataTable({ keySet, headers, items, setItems, setDelete, btnMsg }) {
           {items.map((item, index) => (
             <S.TableTr key={`${keySet}_${index}`}>
               <S.TableTd>
-                <input
-                  type="checkbox"
+                <CheckBox
                   onChange={() => onChecked(item)}
                   checked={selectedLists.has(index)}
                 />
